move add problem type dialog
add parent problem type in list view
