+++ src/app/components/application-areas/application-areas-list/application-areas-list.component.html
@@ -7,6 +7,7 @@
                [allowSelection]=true
                [allowSearch]=true
                [allowSort]=true
+               [allowEdit]=true
                [pagination]=pagingInfo
                [paginatorConfig]=paginatorConfig
                [emptyTableMessage]="'No application areas types found'"
