+++ src/app/components/problem-types/problem-types-list/problem-types-list.component.html
@@ -7,12 +7,14 @@
                [allowSelection]=true
                [allowSearch]=true
                [allowSort]=true
+               [allowEdit]=true
                [pagination]=pagingInfo
                [paginatorConfig]=paginatorConfig
                [emptyTableMessage]="'No problem types found'"
                (elementClicked)="onElementClicked($event)"
                (addElement)="onAddElement()"
                (submitSelectedElements)="onDeleteElements($event)"
+               (updateClicked)="onEditElement($event)"
                (datalistConfigChanged)="getProblemTypes($event)"
                (pageChange)="getProblemTypesHateoas($event)">
     </app-table>
